requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5
annotated-types==0.7.0
torch==2.9.1
numpy==2.2.6
transformers==4.57.1
sentence-transformers==5.1.2
huggingface-hub==0.36.0
google-genai==1.50.1
google-ai-generativelanguage==0.9.0
grpcio==1.76.0
langchain-chroma==1.0.0
langchain-huggingface==1.0.1
langchain-google-genai==3.2.0
langchain-core==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
python-multipart==0.0.21
dependency-injector==4.40.0